pyproject.toml
[tool.poetry]
name = "blog-platform"
version = "0.1.0"
description = ""
authors = ["Azizjon <azizaliev2337@gmail.com>"]
readme = "README.md"
packages = [{include = "blog_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.2.0"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.4"
gunicorn = "^21.2.0"
pytest = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
